requirements.txt
# Core Dependencies
psutil>=5.9.0
requests>=2.28.0
lxml>=4.9.1
pyOpenSSL>=22.1.0
cryptography>=38.0.0

# Windows-specific dependencies
pywin32>=304; platform_system=="Windows"
wmi>=1.5.1; platform_system=="Windows"